break out exp/views/study.py into multiple shorter topical files; add lab update view
