+++ exp/urls.py
@@ -21,6 +21,7 @@ from django.views.decorators.csrf import csrf_exempt
 from exp.views import (
     ExperimenterDashboardView,
     LabDetailView,
+    LabUpdateView,
     ParticipantDetailView,
     ParticipantListView,
     PreviewProxyView,
@@ -58,6 +59,7 @@ app_name = "exp"
 
 urlpatterns = [
     path("labs/<int:pk>/", LabDetailView.as_view(), name="lab-detail"),
+    path("labs/<int:pk>/edit/", LabUpdateView.as_view(), name="lab-edit"),
     path("researchers/", ResearcherListView.as_view(), name="researcher-list"),
     path(
         "researchers/<int:pk>/",
