+++ exp/views/__init__.py
@@ -2,6 +2,14 @@ from exp.views.user import *  # noqa
 
 from exp.views.study import *  # noqa
 
+from exp.views.lab import *  # noqa
+
+from exp.views.responses import *  # noqa
+
+from exp.views.contact import *  # noqa
+
+from exp.views.analytics import *  # noqa
+
 from exp.views.dashboard import *  # noqa
 
 from exp.views.video import *  # noqa
