+++ studies/templates/studies/lab_update.html
@@ -1,0 +1,52 @@
+{% extends 'exp/base.html' %}
+{% load bootstrap3 %}
+
+
+{% block title %}Update lab{% endblock %}
+
+{% block content %}
+
+
+  <div class="container">
+    <div class="row">
+        <div class="col-xs-12">
+            <ol class="breadcrumb">
+              <li><a href=""> Labs </a></li> <!--TODO-->
+                <li><a href="{% url 'exp:lab-detail' pk=lab.id %}"> {{lab.name}} </a></li>
+              <li class="active">Update Lab</li>
+            </ol>
+        </div>
+    </div>
+    <div class="row">
+        <div class="col-xs-12 col-xs-offset-0 col-md-10 col-md-offset-1" id="study-create-form">
+            <div class="panel panel-default mt-md">
+              <div class="panel-heading">
+                <h1 class="panel-title">Update Lab</h1>
+              </div>
+              <div class="panel-body">
+                <form action="" method="post">
+                    {% csrf_token %}
+                    {% bootstrap_form_errors form %}
+                    {% bootstrap_field form.name %}
+                    {% bootstrap_field form.institution %}
+                    {% bootstrap_field form.principal_investigator_name %}
+                    {% bootstrap_field form.contact_email %}
+                    {% bootstrap_field form.contact_phone %}
+                    {% bootstrap_field form.lab_website %}
+                    {% bootstrap_field form.description %}
+                    {% bootstrap_field form.irb_contact_info %}
+                    <div class="modal-footer"> {% csrf_token %}
+                        <a class="btn btn-default"
+                           href="{% url 'exp:lab-edit' pk=lab.id %}"> Discard Changes </a>
+                        <button id="save_study_details_confirm" type="submit"
+                                class="btn btn-danger">
+                            Save
+                        </button>
+                    </div>
+                </form>
+            </div>
+          </div>
+        </div>
+    </div>
+  </div>
+{% endblock %}
